requirements_f.txt
backcall==0.2.0
certifi==2020.12.5
click==7.1.2
cycler==0.10.0
Cython==0.29.21
decorator==4.4.2
gensim==3.8.3
h5py 
ipykernel 
ipython 
ipython-genutils==0.2.0
jedi 
joblib 
jupyter-client 
jupyter-core==4.6.3
kiwisolver==1.2.0
matplotlib 
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
nltk 
notify2==0.3.1
numpy 
olefile==0.46
pandas 
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow 
portalocker==2.0.0
prompt-toolkit 
ptyprocess==0.6.0
pycocotools==2.0.2
Pygments 
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
pyzmq==19.0.2
regex 
sacrebleu==1.4.14
scikit-learn 
scipy 
sentencepiece==0.1.94
sip==4.19.13
six==1.15.0
smart-open==4.0.1
threadpoolctl 
torch==1.5.1
torchsummary==1.5.1
torchvision==0.6.0a0+35d732a
tornado==6.0.4
tqdm 
traitlets 
wcwidth 
